+++ doc/conf.py
@@ -31,16 +31,7 @@
 # ones.
 
 from __future__ import print_function
-import os
 
-
-if not os.path.exists('changelog.rst') and os.path.exists('../CHANGES.rst'):
-    print('Linking ../CHANGES.rst to changelog.rst')
-    if hasattr(os, 'symlink'):
-        os.symlink('../CHANGES.rst', 'changelog.rst')
-    else:
-        import shutil
-        shutil.copyfile('../CHANGES.rst', 'changelog.rst')
 
 extensions = [
     'sphinx.ext.autodoc',