+++ src/nti/schema/fieldproperty.py
@@ -3,10 +3,9 @@
 """
 Computed attributes based on schema fields.
 
-.. $Id$
 """
 
-from __future__ import print_function, unicode_literals, absolute_import, division
+from __future__ import print_function, absolute_import, division
 __docformat__ = "restructuredtext en"
 
 logger = __import__('logging').getLogger(__name__)