+++ CHANGES.rst
@@ -5,10 +5,16 @@
 =========
 
 
-1.1.4 (unreleased)
+1.2.0 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Remove use of ``unicode_literals``.
+
+- Add support for Python 3.6.
+
+- The ``SchemaConfigured`` constructor doesn't hide errors when
+  checking for properties on Python 2. See `issue 11
+  <https://github.com/NextThought/nti.schema/issues/11>`_.
 
 
 1.1.3 (2017-01-17)